refactor(cli): split node fallback dispatchers
