+++ packages/wp-typia/src/node-cli.ts
@@ -10,16 +10,9 @@ import {
   ADD_OPTION_METADATA,
   buildCommandOptionParser,
   CREATE_OPTION_METADATA,
-  type CommandOptionMetadataMap,
-  DOCTOR_OPTION_METADATA,
   extractKnownOptionValuesFromArgv,
-  formatNodeFallbackOptionHelp,
-  INIT_OPTION_METADATA,
-  MIGRATE_OPTION_METADATA,
   parseCommandArgvWithMetadata,
   resolveCommandOptionValues,
-  SYNC_OPTION_METADATA,
-  TEMPLATES_OPTION_METADATA,
 } from './command-option-metadata';
 import { prefersStructuredCliArgv } from './cli-diagnostic-output';
 import {
@@ -31,10 +24,6 @@ import {
   listTemplates,
 } from '@wp-typia/project-tools/cli-templates';
 import {
-  formatAddKindList,
-  formatAddKindUsagePlaceholder,
-} from './add-kind-registry';
-import {
   getAddBlockDefaults,
   getCreateDefaults,
   loadWpTypiaUserConfig,
@@ -44,8 +33,6 @@ import {
 import { extractWpTypiaConfigOverride } from './config-override';
 import type { PrintLine } from './print-line';
 import {
-  executeAddCommand,
-  executeCreateCommand,
   executeDoctorCommand,
   executeInitCommand,
   executeMigrateCommand,
@@ -54,66 +41,35 @@ import {
 } from './runtime-bridge';
 import {
   buildStructuredInitSuccessPayload,
-  buildStructuredCompletionSuccessPayload,
   buildSyncDryRunPayload,
-  extractCompletionProjectDir,
   printCompletionPayload,
 } from './runtime-bridge-output';
 import { resolveSyncExecutionTarget } from './runtime-bridge-sync';
 import {
-  WP_TYPIA_CANONICAL_CREATE_USAGE,
-  WP_TYPIA_CANONICAL_MIGRATE_USAGE,
-  WP_TYPIA_FUTURE_COMMAND_TREE,
-  WP_TYPIA_NODE_FALLBACK_TOP_LEVEL_COMMAND_NAMES,
-  WP_TYPIA_POSITIONAL_ALIAS_USAGE,
   normalizeWpTypiaArgv,
   resolveCanonicalCommandContext,
 } from './command-contract';
-
-type GlobalFlags = {
-  format?: string;
-  id?: string;
-};
-type NodeFallbackTopLevelCommandName =
-  (typeof WP_TYPIA_NODE_FALLBACK_TOP_LEVEL_COMMAND_NAMES)[number];
-type NodeFallbackExecutableCommandName = Exclude<
-  NodeFallbackTopLevelCommandName,
-  'help' | 'version'
->;
-type NodeFallbackDispatchContext = {
-  cwd: string;
-  mergedFlags: Record<string, unknown>;
-  positionals: string[];
-};
-type NodeFallbackCommandHelpConfig = {
-  bodyLines?: string[];
-  heading: string;
-  optionMetadata: CommandOptionMetadataMap;
-};
+import { dispatchNodeFallbackAdd } from './node-fallback/dispatchers/add';
+import { dispatchNodeFallbackCreate } from './node-fallback/dispatchers/create';
+import {
+  NODE_FALLBACK_HELP_RENDERERS,
+  STANDALONE_GUIDANCE_LINE,
+  renderGeneralHelp,
+} from './node-fallback/help';
+import type {
+  NodeFallbackCommandDispatcher,
+  NodeFallbackDispatchContext,
+  NodeFallbackExecutableCommandName,
+  NodeFallbackGlobalFlags,
+} from './node-fallback/types';
 
 const NODE_FALLBACK_OPTION_PARSER = buildCommandOptionParser(
   ALL_COMMAND_OPTION_METADATA,
 );
 const NODE_FALLBACK_BOOLEAN_OPTION_NAMES = ['help', 'version'] as const;
-const STANDALONE_GUIDANCE_LINE =
-  'Prefer not to install Bun? Use the standalone wp-typia binary from the GitHub release assets.';
-
-const NODE_FALLBACK_RUNTIME_SUMMARY_LINES = [
-  'Runtime: Node fallback',
-  'Human-readable fallback for common non-interactive create/init/add/migrate flows, doctor, sync, templates, --help, and --version when Bun is unavailable.',
-  `Install Bun 1.3.11+ or use \`bunx wp-typia ...\` for the full Bunli/OpenTUI runtime and Bun-only command surfaces such as \`skills\`, \`completions\`, and \`mcp\`. ${STANDALONE_GUIDANCE_LINE}`,
-  'Output markers: WP_TYPIA_ASCII=1 forces ASCII markers, WP_TYPIA_ASCII=0 opts back into Unicode markers, and non-empty NO_COLOR requests ASCII markers when WP_TYPIA_ASCII is unset.',
-];
-
 const printLine: PrintLine = (line) => {
   console.log(line);
 };
-
-function printBlock(lines: string[]) {
-  for (const line of lines) {
-    printLine(line);
-  }
-}
 
 export function hasFlagBeforeTerminator(argv: string[], flag: string): boolean {
   for (const arg of argv) {
@@ -130,7 +86,7 @@ export function hasFlagBeforeTerminator(argv: string[], flag: string): boolean {
 
 export function parseGlobalFlags(argv: string[]): {
   argv: string[];
-  flags: GlobalFlags;
+  flags: NodeFallbackGlobalFlags;
 } {
   const { argv: nextArgv, flags } = extractKnownOptionValuesFromArgv(argv, {
     optionNames: ['format', 'id'],
@@ -192,87 +148,6 @@ function parseArgv(argv: string[]) {
   });
 }
 
-function renderGeneralHelp() {
-  printBlock([
-    `wp-typia ${packageJson.version}`,
-    '',
-    'Canonical CLI package for wp-typia scaffolding and project workflows.',
-    '',
-    ...NODE_FALLBACK_RUNTIME_SUMMARY_LINES,
-    '',
-    'Commands:',
-    ...WP_TYPIA_FUTURE_COMMAND_TREE.map(
-      (command) => `- ${command.name}: ${command.description}`,
-    ),
-    '',
-    'Canonical usage:',
-    `- ${WP_TYPIA_CANONICAL_CREATE_USAGE}`,
-    '- wp-typia init [project-dir]',
-    `- ${WP_TYPIA_CANONICAL_MIGRATE_USAGE}`,
-    `- ${WP_TYPIA_POSITIONAL_ALIAS_USAGE}`,
-  ]);
-}
-
-/**
- * Render one Node fallback command help screen from shared command metadata.
- */
-function renderNodeFallbackCommandHelp(config: NodeFallbackCommandHelpConfig) {
-  printBlock([
-    config.heading,
-    '',
-    ...NODE_FALLBACK_RUNTIME_SUMMARY_LINES,
-    '',
-    ...(config.bodyLines ? [...config.bodyLines, ''] : []),
-    'Supported flags:',
-    ...formatNodeFallbackOptionHelp(config.optionMetadata),
-  ]);
-}
-
-const NODE_FALLBACK_COMMAND_HELP_CONFIG = {
-  add: {
-    bodyLines: [`Supported kinds: ${formatAddKindList()}`],
-    heading: 'Usage: wp-typia add <kind> <name>',
-    optionMetadata: ADD_OPTION_METADATA,
-  },
-  create: {
-    heading: `Usage: ${WP_TYPIA_CANONICAL_CREATE_USAGE}`,
-    optionMetadata: CREATE_OPTION_METADATA,
-  },
-  doctor: {
-    bodyLines: [
-      'Runs read-only environment readiness checks. Official wp-typia workspace roots also get inventory, source-tree drift, iframe/API v3 compatibility, and shared convention checks.',
-    ],
-    heading: 'Usage: wp-typia doctor [--format json]',
-    optionMetadata: DOCTOR_OPTION_METADATA,
-  },
-  init: {
-    bodyLines: [
-      'Preview-by-default retrofit planner for existing WordPress block or plugin projects. Re-run with --apply to write package.json updates and helper scripts.',
-    ],
-    heading: 'Usage: wp-typia init [project-dir]',
-    optionMetadata: INIT_OPTION_METADATA,
-  },
-  migrate: {
-    heading: `Usage: ${WP_TYPIA_CANONICAL_MIGRATE_USAGE}`,
-    optionMetadata: MIGRATE_OPTION_METADATA,
-  },
-  sync: {
-    heading: 'Usage: wp-typia sync [ai]',
-    optionMetadata: SYNC_OPTION_METADATA,
-  },
-  templates: {
-    heading: 'wp-typia templates <list|inspect>',
-    optionMetadata: TEMPLATES_OPTION_METADATA,
-  },
-} satisfies Record<NodeFallbackExecutableCommandName, NodeFallbackCommandHelpConfig>;
-
-const NODE_FALLBACK_HELP_RENDERERS = Object.fromEntries(
-  Object.entries(NODE_FALLBACK_COMMAND_HELP_CONFIG).map(([command, config]) => [
-    command,
-    () => renderNodeFallbackCommandHelp(config),
-  ]),
-) as Record<NodeFallbackExecutableCommandName, () => void>;
-
 function renderVersion(
   options: {
     format?: string;
@@ -299,7 +174,10 @@ function renderVersion(
   printLine(`wp-typia ${packageJson.version}`);
 }
 
-function renderTemplatesJson(flags: GlobalFlags, subcommand: string) {
+function renderTemplatesJson(
+  flags: NodeFallbackGlobalFlags,
+  subcommand: string,
+) {
   if (subcommand === 'list') {
     printLine(
       JSON.stringify(
@@ -384,110 +262,8 @@ async function renderDoctorJson(): Promise<void> {
 }
 
 const NODE_FALLBACK_COMMAND_DISPATCHERS = {
-  add: async ({
-    cwd,
-    mergedFlags,
-    positionals,
-  }: NodeFallbackDispatchContext) => {
-    if (!positionals[1]) {
-      const { formatAddHelpText } =
-        await import('@wp-typia/project-tools/cli-add');
-      printLine(formatAddHelpText());
-      throw createCliCommandError({
-        code: CLI_DIAGNOSTIC_CODES.MISSING_ARGUMENT,
-        command: 'add',
-        detailLines: [
-          `\`wp-typia add\` requires <kind>. Usage: wp-typia add ${formatAddKindUsagePlaceholder()} ...`,
-        ],
-      });
-    }
-    if (mergedFlags.format === 'json') {
-      let completion;
-      try {
-        completion = await executeAddCommand({
-          cwd,
-          emitOutput: false,
-          flags: mergedFlags,
-          interactive: false,
-          kind: positionals[1],
-          name: positionals[2],
-        });
-      } catch (error) {
-        throw createCliCommandError({
-          command: 'add',
-          error,
-        });
-      }
-      printLine(
-        JSON.stringify(
-          buildStructuredCompletionSuccessPayload('add', completion, {
-            dryRun: Boolean(mergedFlags['dry-run']),
-            kind: positionals[1],
-            name: positionals[2],
-            projectDir: extractCompletionProjectDir(completion) ?? cwd,
-          }),
-          null,
-          2,
-        ),
-      );
-      return;
-    }
-    await executeAddCommand({
-      cwd,
-      flags: mergedFlags,
-      interactive: undefined,
-      kind: positionals[1],
-      name: positionals[2],
-    });
-  },
-  create: async ({
-    cwd,
-    mergedFlags,
-    positionals,
-  }: NodeFallbackDispatchContext) => {
-    const projectDir = positionals[1];
-    if (!projectDir) {
-      throw createCliCommandError({
-        code: CLI_DIAGNOSTIC_CODES.MISSING_ARGUMENT,
-        command: 'create',
-        detailLines: [
-          '`wp-typia create` requires <project-dir>.',
-          '`--dry-run` still needs a logical project directory name because wp-typia derives slugs, package names, and planned file paths from it.',
-        ],
-      });
-    }
-    let completion;
-    try {
-      completion = await executeCreateCommand({
-        cwd,
-        emitOutput: mergedFlags.format !== 'json',
-        flags: mergedFlags,
-        interactive: mergedFlags.format === 'json' ? false : undefined,
-        projectDir,
-      });
-    } catch (error) {
-      throw createCliCommandError({
-        command: 'create',
-        error,
-      });
-    }
-    if (mergedFlags.format === 'json') {
-      printLine(
-        JSON.stringify(
-          buildStructuredCompletionSuccessPayload('create', completion, {
-            dryRun: Boolean(mergedFlags['dry-run']),
-            projectDir: extractCompletionProjectDir(completion) ?? projectDir,
-            template:
-              typeof mergedFlags.template === 'string'
-                ? mergedFlags.template
-                : undefined,
-          }),
-          null,
-          2,
-        ),
-      );
-    }
-  },
+  add: dispatchNodeFallbackAdd,
+  create: dispatchNodeFallbackCreate,
   doctor: async ({ cwd, mergedFlags }: NodeFallbackDispatchContext) => {
     if (mergedFlags.format === 'json') {
       await renderDoctorJson();
@@ -614,7 +390,7 @@ const NODE_FALLBACK_COMMAND_DISPATCHERS = {
   },
 } satisfies Record<
   NodeFallbackExecutableCommandName,
-  (context: NodeFallbackDispatchContext) => Promise<void>
+  NodeFallbackCommandDispatcher
 >;
 
 export async function runNodeCli(argv = process.argv.slice(2)): Promise<void> {
@@ -639,7 +415,7 @@ export async function runNodeCli(argv = process.argv.slice(2)): Promise<void> {
     hasFlagBeforeTerminator(cliArgv, '--version') || command === 'version';
 
   if (cliArgv.length === 0) {
-    renderGeneralHelp();
+    renderGeneralHelp(printLine);
     process.exitCode = 1;
     return;
   }
@@ -651,18 +427,18 @@ export async function runNodeCli(argv = process.argv.slice(2)): Promise<void> {
           helpTarget as NodeFallbackExecutableCommandName
         ];
       if (helpRenderer) {
-        helpRenderer();
+        helpRenderer(printLine);
         return;
       }
       if (helpTarget === 'help' || helpTarget === 'version') {
-        renderGeneralHelp();
+        renderGeneralHelp(printLine);
         return;
       }
     } else {
-      renderGeneralHelp();
+      renderGeneralHelp(printLine);
       return;
     }
-    renderGeneralHelp();
+    renderGeneralHelp(printLine);
     return;
   }
 
@@ -694,6 +470,7 @@ export async function runNodeCli(argv = process.argv.slice(2)): Promise<void> {
       cwd: process.cwd(),
       mergedFlags,
       positionals,
+      printLine,
     });
     return;
   }
