+++ packages/wp-typia/src/node-fallback/dispatchers/create.ts
@@ -1,0 +1,64 @@
+import {
+  CLI_DIAGNOSTIC_CODES,
+  createCliCommandError,
+} from '@wp-typia/project-tools/cli-diagnostics';
+import { executeCreateCommand } from '../../runtime-bridge';
+import {
+  buildStructuredCompletionSuccessPayload,
+  extractCompletionProjectDir,
+} from '../../runtime-bridge-output';
+import type { NodeFallbackDispatchContext } from '../types';
+
+export async function dispatchNodeFallbackCreate({
+  cwd,
+  mergedFlags,
+  positionals,
+  printLine,
+}: NodeFallbackDispatchContext): Promise<void> {
+  const projectDir = positionals[1];
+  if (!projectDir) {
+    throw createCliCommandError({
+      code: CLI_DIAGNOSTIC_CODES.MISSING_ARGUMENT,
+      command: 'create',
+      detailLines: [
+        '`wp-typia create` requires <project-dir>.',
+        '`--dry-run` still needs a logical project directory name because wp-typia derives slugs, package names, and planned file paths from it.',
+      ],
+    });
+  }
+
+  // Create-specific normalization stays here: JSON mode suppresses human output
+  // and forces non-interactive execution after shared flag parsing/defaults.
+  let completion;
+  try {
+    completion = await executeCreateCommand({
+      cwd,
+      emitOutput: mergedFlags.format !== 'json',
+      flags: mergedFlags,
+      interactive: mergedFlags.format === 'json' ? false : undefined,
+      projectDir,
+    });
+  } catch (error) {
+    throw createCliCommandError({
+      command: 'create',
+      error,
+    });
+  }
+
+  if (mergedFlags.format === 'json') {
+    printLine(
+      JSON.stringify(
+        buildStructuredCompletionSuccessPayload('create', completion, {
+          dryRun: Boolean(mergedFlags['dry-run']),
+          projectDir: extractCompletionProjectDir(completion) ?? projectDir,
+          template:
+            typeof mergedFlags.template === 'string'
+              ? mergedFlags.template
+              : undefined,
+        }),
+        null,
+        2,
+      ),
+    );
+  }
+}
