+++ packages/wp-typia/src/node-fallback/types.ts
@@ -1,0 +1,26 @@
+import type { WP_TYPIA_NODE_FALLBACK_TOP_LEVEL_COMMAND_NAMES } from '../command-contract';
+import type { PrintLine } from '../print-line';
+
+export type NodeFallbackGlobalFlags = {
+  format?: string;
+  id?: string;
+};
+
+export type NodeFallbackTopLevelCommandName =
+  (typeof WP_TYPIA_NODE_FALLBACK_TOP_LEVEL_COMMAND_NAMES)[number];
+
+export type NodeFallbackExecutableCommandName = Exclude<
+  NodeFallbackTopLevelCommandName,
+  'help' | 'version'
+>;
+
+export type NodeFallbackDispatchContext = {
+  cwd: string;
+  mergedFlags: Record<string, unknown>;
+  positionals: string[];
+  printLine: PrintLine;
+};
+
+export type NodeFallbackCommandDispatcher = (
+  context: NodeFallbackDispatchContext,
+) => Promise<void>;
